File structure, Work, Images, JS
Updated backgrounds to be the right size [Hero, Work, Socials] along with setting the min-height on these sections to be 1080px. Added pictures and videos of renders and other work to work section and small slideshow at top. Added art section to the work area. Modified file structure to split work and favicons to be separate. Removed previous placeholder images and replaced them with the CCS Design Department one where required. Updated titles, descriptions and captions to work where applicable.  Added folder "Non-Live_Sites" for websites i don't want to host on a new subdomain but show on the portfolio. Added the ability to download for the 360 video along with making it a case in the render if statement in the css along with one for videos and removing the one for the iframe test. Shrunk the instagram iframe on socials area to make it fit better with the new size.